requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
emoji==0.5.4
future==0.18.2
pkg-resources==0.0.0
pycparser==2.20
python-telegram-bot==12.7
six==1.14.0
tornado==6.0.4
clarifai==2.6.2
pymongo-3.10.1
dnspython-1.16.0